feat:Employee and Product CRUD with pattern

+++ src/app/dashboard/employee/list-employee/list-employee.component.ts
@@ -1,5 +1,10 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { FormGroup } from '@angular/forms';
+import { MatButton } from '@angular/material/button';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { AddEditEmployeeComponent } from '../add-edit-employee/add-edit-employee.component';
+import { DialogService } from '../dialog.service';
 import { Employee } from '../employee';
 import { EmployeeService } from '../employee.service';
 
@@ -10,39 +15,69 @@ import { EmployeeService } from '../employee.service';
 })
 export class ListEmployeeComponent implements OnInit {
   public employee!: Employee[];
+  public employeeFormGroup !: FormGroup;
+  public searchText: string = '';
   @Output() editClick: EventEmitter<string> = new EventEmitter<string>();
 
-  constructor(private service: EmployeeService, private router: Router,
-    ) {
-    this.getEmployees();
-   }
+  constructor(private service: EmployeeService,
+             private router: Router, 
+             public dialog: MatDialog,
+             public dialogService: DialogService  ) {
+  
+      this.getEmployees();
+
+    }
+
+  public addEmp(): void {
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.width = "35%";
+    this.dialog.open(AddEditEmployeeComponent, dialogConfig);
+  }
 
   ngOnInit(): void {
+    
   }
 
   //Get All Employee detail
   public getEmployees() {
     return this.service.getEmployees().subscribe((data: Employee[]) => {
-      
       this.employee = data;
     })
   }
 
+
   //Get Employee by Id from Rest API
-  public editEmp(id:number){
-    //this.editClick.emit(`${id}`);
-    this.service.getEmployeeById(id).subscribe((data: Employee) => {      
-      alert("Employee has been edited from list");
-      this.router.navigate([`../dashboard/Employee/addEdit/${id}`]);
-    })  
+  public editEmp(emp: Employee) {
+    this.service.getEmployeeById(emp.id).subscribe((data: Employee) => {
+
+      const dialogConfig = new MatDialogConfig();
+      dialogConfig.disableClose = true;
+      dialogConfig.data = emp;
+      dialogConfig.autoFocus = true;
+      dialogConfig.width = "35%";
+      this.dialog.open(AddEditEmployeeComponent, dialogConfig);   
+    })
   }
 
+
   //Delete Employee details
-  public deleteEmp(id:number){
-     this.service.deleteEmployee(id).subscribe((data: number) => {
-        alert("Employee deleted successfully");
-        this.getEmployees();
-      })
+  public deleteEmp(id: number) {
+    debugger
+    this.dialogService.openConfirmDialog("Are you sure want to delete this record ?")
+    .afterClosed().subscribe(res =>{
+      if(res){
+        this.service.deleteEmployee(id).subscribe((data:number)=>{
+          this.getEmployees();
+        })
+      }
+    });
+    // this.service.deleteEmployee(id).subscribe((data: number) => {
+    //   alert('Employee deleted successfully');
+      
   }
+
+
 
 }
